Remove verbosity from Ansible deployment
After activating the production branch trigger in the .drone.yml
configuration file, the automatic deployment to production went
smoothly. It looks like the previous error on the production branch was
caused by the deactivated production branch and stale data on drone CI.

This commit removes the SSH and Ansible internals verbosity.