Improve JSON Linting
The JSON linting should be done by the `jsonlint` CLI. This commit
changes the CI command to fail when JSONLint encounters an error.
Previously it only output the error.

Deleting the `JsonIsValidTest` because it did the same, but did not have
the detailed error message about *what* is wrong with the JSON and which
file has an error.