Do not trigger deploy on pull requests
When doing a pull request from a feature branch on GitHub, the `branch`
variable is the *target* of the pull request not the name of the feature
branch.
See https://woodpecker-ci.org/docs/usage/workflow-syntax#branch

This commit removes the `pull_request` condition, which means that only
pushes to the `test` and `production` branches will trigger the
deployment.

Ticket: https://phabricator.wikimedia.org/T396226